requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
flask
requests
shap
google-generativeai
